Optimization of wordsNearestSum
After performing the multiplication, sorting the whole matrix and then extracting the top N words is expensive. I switched the implementation to a priority queue. The result is a boost in the performance. For the word `France` in the `GoogleNews` corpus, the original implementation takes `2908 ms` in sorting. The new implementation takes only `102 ms` to perform the extraction.